requirements.txt
aiofile==3.3.3
aiofiles==0.6.0      
aiohttp==3.6.3       
aiosqlite==0.17.0       
async-timeout==3.0.1 
certifi==2020.12.5   
discord.py==1.5.1
distlib==0.3.1
facebook-scraper==0.2.35
lyricsgenius==3.0.1
matplotlib==3.3.3
nodeenv==1.6.0
numpy==1.20.0
pandas==1.2.2
parse==1.18.0
parso==0.8.1
pathlib==1.0.1
pickleshare==0.7.5
Pillow==8.0.1
psutil==5.8.0
PyNaCl==1.3.0
pypng==0.0.20
PyQRCode==1.2.1
python-dotenv==0.14.0
pytube==10.6.1
regex==2021.3.17
requests==2.25.1
sympy==1.7.1
timeago==1.0.13
translate==3.5.0
typed-dotenv==1.0.1
urllib3==1.26.2
virtualenv==20.4.4
wikipedia==1.4.0
instaloader==4.7.1
jokeapi==0.2.11
discord-py-slash-command==1.2.2
discord-components==0.5.3
asyncpraw==7.2.0
asyncprawcore==2.0.1
update-checker==0.18.0